Only restrict settings to admins.
